requirements.txt
# Oracle DBA RAG Agent — Python Dependencies
# Python 3.11/3.12/3.13 recommended. Python 3.14 NOT yet supported by grpcio.
# Install: pip install -r requirements.txt

# ── Core RAG Pipeline ─────────────────────────────────────────
sentence-transformers>=2.7.0,<3.0.0
# qdrant-client: use REST-only mode to avoid grpcio build issues on Python 3.14
qdrant-client[rest]>=1.9.1,<2.0.0
rank-bm25==0.2.2
duckdb>=0.10.3,<2.0.0
networkx>=3.3,<4.0.0

# ── Log Parsing ───────────────────────────────────────────────
python-dateutil>=2.9.0
pytz>=2024.1
regex>=2024.4.28

# ── API Layer ─────────────────────────────────────────────────
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic>=2.7.1,<3.0.0
python-multipart>=0.0.9

# ── UI ────────────────────────────────────────────────────────
streamlit>=1.34.0,<2.0.0

# ── Utilities ─────────────────────────────────────────────────
PyYAML>=6.0.1
pypdf>=4.0.0,<6.0.0
python-dotenv>=1.0.1
rich>=13.7.1
tqdm>=4.66.4
loguru>=0.7.2
numpy>=1.26.4,<3.0.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.6
